+++ service/src/main/kotlin/io/provenance/explorer/grpc/extensions/Domain.kt
@@ -1,4 +1,4 @@
-package io.provenance.explorer.grpc
+package io.provenance.explorer.grpc.extensions
 
 import com.google.protobuf.Any
 import com.google.protobuf.ByteString